# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Donut Market Terminal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /items/{scopeKey}:
    get:
      operationId: getItemDetail
      tags: [items]
      summary: Item detail
      parameters:
        - { name: scopeKey, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /items/history:
    get:
      operationId: getItemHistory
      tags: [items]
      summary: Item OHLC price history
      parameters:
        - { name: scopeKey, in: query, required: true, schema: { type: string } }
        - { name: interval, in: query, schema: { type: string, enum: ["5m", "1h", "1d"], default: "1h" } }
        - { name: range, in: query, schema: { type: string, enum: ["1h", "6h", "24h", "7d", "30d", "90d", "all"], default: "7d" } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceHistory"
  /items/{scopeKey}/listings:
    get:
      operationId: getItemListings
      tags: [items]
      summary: Current listings for an item
      parameters:
        - { name: scopeKey, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Listing" }
  /items/transactions:
    get:
      operationId: getItemTransactions
      tags: [items]
      summary: Recent sales for an item
      parameters:
        - { name: scopeKey, in: query, required: true, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 50 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Sale" }
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    Enchant:
      type: object
      required: [name, level]
      properties:
        name: { type: string }
        level: { type: integer }
    Listing:
      type: object
      required: [id, scopeKey, baseItemId, displayName, quantity, totalPrice, unitPrice]
      properties:
        id: { type: integer }
        scopeKey: { type: string }
        variantScopeKey: { type: string }
        baseItemId: { type: string }
        displayName: { type: string }
        sellerName: { type: ["string", "null"] }
        sellerUuid: { type: ["string", "null"] }
        quantity: { type: integer }
        totalPrice: { type: number }
        unitPrice: { type: number }
        timeLeftMs: { type: ["number", "null"] }
        approxExpiresAt: { type: ["string", "null"] }
        enchants:
          type: array
          items: { $ref: "#/components/schemas/Enchant" }
        trimMaterial: { type: ["string", "null"] }
        trimPattern: { type: ["string", "null"] }
        lore:
          type: array
          items: { type: string }
...
    Sale:
      type: object
      required: [id, scopeKey, baseItemId, displayName, quantity, totalPrice, unitPrice, soldAt]
      properties:
        id: { type: integer }
        scopeKey: { type: string }
        variantScopeKey: { type: string }
        baseItemId: { type: string }
        displayName: { type: string }
        sellerName: { type: ["string", "null"] }
        sellerUuid: { type: ["string", "null"] }
        quantity: { type: integer }
        totalPrice: { type: number }
        unitPrice: { type: number }
        soldAt: { type: string }
        enchants:
          type: array
          items: { $ref: "#/components/schemas/Enchant" }
...
    MarketSignal:
      type: object
      required: [label, detail]
      properties:
        label: { type: string }
        detail: { type: string }
        tone: { type: string }
    ItemMetrics:
      type: object
      required: [confidence]
      properties:
        latestSale: { type: ["number", "null"] }
        bestAsk: { type: ["number", "null"] }
        medianAsk: { type: ["number", "null"] }
        median1h: { type: ["number", "null"] }
        median24h: { type: ["number", "null"] }
        median7d: { type: ["number", "null"] }
        median30d: { type: ["number", "null"] }
        change24h: { type: ["number", "null"] }
        change7d: { type: ["number", "null"] }
        change30d: { type: ["number", "null"] }
        salesCount24h: { type: integer }
        soldQty24h: { type: number }
        tradedValue24h: { type: number }
        activeListings: { type: integer }
        listedQty: { type: number }
        velocityPerHour: { type: ["number", "null"] }
        velocityPerDay: { type: ["number", "null"] }
        listingToSalesRatio: { type: ["number", "null"] }
        daysOfSupply: { type: ["number", "null"] }
        volatility: { type: ["number", "null"] }
        askPremiumPct: { type: ["number", "null"] }
        high24h: { type: ["number", "null"] }
        low24h: { type: ["number", "null"] }
        confidence: { type: number }
    ItemVariantInfo:
      type: object
      required: [scopeKey, baseItemId, displayName]
      properties:
        scopeKey: { type: string }
        baseItemId: { type: string }
        displayName: { type: string }
        variantHash: { type: ["string", "null"] }
        enchants:
          type: array
          items: { $ref: "#/components/schemas/Enchant" }
        trimMaterial: { type: ["string", "null"] }
        trimPattern: { type: ["string", "null"] }
        lore:
          type: array
          items: { type: string }
    ItemDetail:
      type: object
      required: [scope, scopeKey, baseItemId, displayName, metrics, signals]
      properties:
        scope: { type: string }
        scopeKey: { type: string }
        baseItemId: { type: string }
        displayName: { type: string }
        baseScopeKey: { type: string }
        mixedVariants: { type: boolean }
        metrics: { $ref: "#/components/schemas/ItemMetrics" }
        signals:
          type: array
          items: { $ref: "#/components/schemas/MarketSignal" }
        variants:
          type: array
          items: { $ref: "#/components/schemas/ItemVariantInfo" }
    Candle:
      type: object
      required: [t]
      properties:
        t: { type: string }
        open: { type: ["number", "null"] }
        high: { type: ["number", "null"] }
        low: { type: ["number", "null"] }
        close: { type: ["number", "null"] }
        median: { type: ["number", "null"] }
        volume: { type: number }
        txCount: { type: integer }
        gap: { type: boolean }
    DepthPoint:
      type: object
      required: [unitPrice, cumulativeQty]
      properties:
        unitPrice: { type: number }
        cumulativeQty: { type: number }
    HistogramBin:
      type: object
      required: [from, to, count]
      properties:
        from: { type: number }
        to: { type: number }
        count: { type: integer }
    PriceHistory:
      type: object
      required: [interval, range, candles, depth, histogram]
      properties:
        interval: { type: string }
        range: { type: string }
        candles:
          type: array
          items: { $ref: "#/components/schemas/Candle" }
        depth:
          type: array
          items: { $ref: "#/components/schemas/DepthPoint" }
        histogram:
          type: array
          items: { $ref: "#/components/schemas/HistogramBin" }
...