# lib/api-zod/src/generated/types/itemMetrics.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Donut Market Terminal API
 * OpenAPI spec version: 0.1.0
 */

export interface ItemMetrics {
  /** @nullable */
  latestSale?: number | null;
  /** @nullable */
  bestAsk?: number | null;
  /** @nullable */
  medianAsk?: number | null;
  /** @nullable */
  median1h?: number | null;
  /** @nullable */
  median24h?: number | null;
  /** @nullable */
  median7d?: number | null;
  /** @nullable */
  median30d?: number | null;
  /** @nullable */
  change24h?: number | null;
  /** @nullable */
  change7d?: number | null;
  /** @nullable */
  change30d?: number | null;
  salesCount24h?: number;
  soldQty24h?: number;
  tradedValue24h?: number;
  activeListings?: number;
  listedQty?: number;
  /** @nullable */
  velocityPerHour?: number | null;
  /** @nullable */
  velocityPerDay?: number | null;
  /** @nullable */
  listingToSalesRatio?: number | null;
  /** @nullable */
  daysOfSupply?: number | null;
  /** @nullable */
  volatility?: number | null;
  /** @nullable */
  askPremiumPct?: number | null;
  /** @nullable */
  high24h?: number | null;
  /** @nullable */
  low24h?: number | null;
  confidence: number;
}